+++ source/renderer/app/Routes.js
@@ -8,6 +8,7 @@ import Root from './containers/Root';
 import InitialSettingsPage from './containers/profile/InitialSettingsPage';
 import Settings from './containers/settings/Settings';
 import GeneralSettingsPage from './containers/settings/categories/GeneralSettingsPage';
+import StakePoolsSettingsPage from './containers/settings/categories/StakePoolsSettingsPage';
 import SupportSettingsPage from './containers/settings/categories/SupportSettingsPage';
 import TermsOfUseSettingsPage from './containers/settings/categories/TermsOfUseSettingsPage';
 import TermsOfUsePage from './containers/profile/TermsOfUsePage';
@@ -88,6 +89,10 @@ export const Routes = withRouter(() => (
             <Route
               path={ROUTES.SETTINGS.GENERAL}
               component={GeneralSettingsPage}
+            />
+            <Route
+              path={ROUTES.SETTINGS.STAKE_POOLS}
+              component={StakePoolsSettingsPage}
             />
             <Route
               path={ROUTES.SETTINGS.TERMS_OF_USE}